fix: resolve Next.js build errors (nullable searchParams, event-bus extraction)
- Extract appearanceGradeEventBus to separate file (invalid Page export)
- Add optional chaining for searchParams/pathname nullable checks
- Build now passes with ignoreBuildErrors: false
