+++ .omc/state/agent-replay-a67174e3-4131-4bc9-8680-89fe8c81a722.jsonl
@@ -5,3 +5,5 @@
 {"t":0,"agent":"a8d8d4f","agent_type":"executor","event":"agent_start","parent_mode":"none"}
 {"t":0,"agent":"aa2d5e4","agent_type":"executor","event":"agent_start","parent_mode":"none"}
 {"t":0,"agent":"aa2d5e4","agent_type":"executor","event":"agent_stop","success":true,"duration_ms":214910}
+{"t":0,"agent":"acb6428","agent_type":"build-fixer","event":"agent_start","parent_mode":"none"}
+{"t":0,"agent":"acb6428","agent_type":"build-fixer","event":"agent_stop","success":true,"duration_ms":81825}
