+++ .omc/state/idle-notif-cooldown.json
@@ -1,3 +1,3 @@
 {
-  "lastSentAt": "2026-03-14T06:49:31.240Z"
+  "lastSentAt": "2026-03-14T06:57:54.122Z"
 }
